% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{doNetwork}
\alias{doNetwork}
\title{Do graph-based internet filtration}
\usage{
doNetwork(dat = NULL, plot = FALSE, outdir = "./",
  outliers.coef = 1.2, debug = FALSE)
}
\arguments{
\item{dat}{_edge.txt file}

\item{plot}{Plot ions-interaction network, default FALSE}

\item{outdir}{Output directory, default current directory}

\item{outliers.coef}{Default 1.2}

\item{debug}{Default FALSE}
}
\value{
MGF
}
\description{
Do graph-based internet filtration
}
