% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{.sel.outliers}
\alias{.sel.outliers}
\title{Remove high-error ion-interactions}
\usage{
.sel.outliers(x, method = "iqr", outliers.coef = 1)
}
\arguments{
\item{x}{Ion-interactions}

\item{method}{iqr method}

\item{outliers.coef}{Default 1.2}
}
\value{
Simplified ion-interactions
}
\description{
Remove high-error ion-interactions
}
