% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mergeMGF}
\alias{mergeMGF}
\title{Merge MGF files}
\usage{
mergeMGF(dir = NULL, name = NULL, clean = TRUE)
}
\arguments{
\item{dir}{The directry of MS/MS data}

\item{name}{Merged file name}

\item{clean}{Delete individual MS/MS spectrum file}
}
\value{
MGF
}
\description{
Merge MGF files
}
\section{Functions}{
\itemize{
\item \code{mergeMGF}: multiply MS/MS data
}}

