% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pCleanGear}
\alias{pCleanGear}
\title{Preprocess high-resolution tandem mass spectra prior to database searching}
\usage{
pCleanGear(mgf = NULL, itol = 0.05, outdir = "./", mem = 1,
  cpu = 0, plot = FALSE, aa2 = TRUE, mionFilter = FALSE,
  labelMethod = NULL, repFilter = FALSE, labelFilter = FALSE,
  low = FALSE, high = FALSE, isoReduction = FALSE,
  chargeDeconv = FALSE, largerThanPrecursor = FALSE,
  ionsMerge = FALSE, network = TRUE, idres = NULL,
  ms2tolfilter = 1.2, debug = FALSE)
}
\arguments{
\item{mgf}{Input MS/MS data in MGF format.}

\item{itol}{Fragment tolerance, default 0.05Da.}

\item{outdir}{Output directory, default current directory.}

\item{mem}{The maximun memeroy to be used, default 1GB.}

\item{cpu}{The number of cpu to be used, default 0 (all).}

\item{aa2}{Consider gap masses of two amino acid, default is TRUE.}

\item{mionFilter}{Filter immonium ions, default FALSE.}

\item{labelMethod}{Peptide labeling method, iTRAQ4plex, iTRAQ8plex, TMT6plex, TMT10plex or NULL.}

\item{repFilter}{Remove reporter ions, default FALSE.}

\item{labelFilter}{Remove label-associated ions, default FALSE.}

\item{low}{Remove ions in low b-/y-free window for label-based MS/MS data, default FALSE.}

\item{high}{Remove ions in high b-/y-free window for label-based MS/MS data, default FALSE.}

\item{isoReduction}{Reduction of heavy isotopic peaks, default FALSE.}

\item{chargeDeconv}{Charge deconvolution, default FALSE.}

\item{largerThanPrecursor}{Remove ions larger than precursor, default FALSE.}

\item{ionsMerge}{Merge two adjecent ions with similar m/z, default FALSE.}

\item{network}{Graph-based network filtration.}

\item{idres}{Use MSGF identification result to annotate peaks (.mzid), default NULL.}

\item{ms2tolfilter}{Default 1.2.}

\item{debug}{Whether to export edge file and peak file.}
}
\value{
A much cleaner MS/MS data.
}
\description{
This package is designed to preprocess high-resolution MS/MS Data, which integrated three modules, removal of label-associated ions, isotope peak reduction and charge deconvolution, and a graph-based network approach and aimed at filtering out extraneous peaks with/without specific-feature. pClean is supportive to a wide array of instruments with all types of MS data, and incorporative into most data analysis pipelines.
}
