% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{simplify}
\alias{simplify}
\title{Simplify the ion-interactions based on the formula level}
\usage{
simplify(x)
}
\arguments{
\item{x}{Ion-interactions}
}
\value{
Simplified ion-interactions
}
\description{
Simplify the ion-interactions based on the formula level
}
